Updated as per the today's JDK 8 API changes.
Item1: Updated to use the the new API which forces to use an IntMapper on an List<Integer>. It now has 2 variants for boxed type and primitive type.
Item2:  Shows another variant that uses the new IntMapper which makes easier to obtain the sum on the IntStream.
Item6 : Fixed a package renaming in the today's build.